requirements.txt
aws-cdk.aws-directoryservice
aws-cdk.aws-ec2
aws-cdk.aws-iam
aws-cdk.core
